requirements.txt
wheel==0.37.1
Django==4.0.2
djangorestframework==3.13.1
coverage==6.3.1
flake8==4.0.1
pyflakes==2.4.0
isort==5.10.1
autopep8==1.6.0
flake8-isort==4.1.1
coreapi==2.3.3
channels==3.0.4
channels_redis==3.3.1
